sys.h から logbuf.{cpp,h} と threading.h を切り出した。
