+++ core/common/sys.h
@@ -21,17 +21,6 @@
 #define _SYS_H
 
 #include <string>
-#include <atomic>
-
-#include <mutex>
-#if defined(__MINGW32__) || defined(__MINGW64__)
-#include "mingw.mutex.h"
-#endif
-
-#if defined(__MINGW32__) || defined(__MINGW64__)
-#include "mingw.thread.h"
-#endif
-#include <thread>
 
 #include <string.h>
 #include <stdarg.h>
@@ -117,181 +106,22 @@ public:
     class LogBuffer *logBuf;
 };
 
-#ifdef WIN32
-#include <windows.h>
-
-typedef __int64 int64_t;
+// ------------------------------------
+extern Sys *sys;
 
 // ------------------------------------
-class WEvent
-{
-public:
-    WEvent()
-    {
-        event = ::CreateEvent(NULL,  // no security attributes
-                              TRUE,  // manual-reset
-                              FALSE, // initially non-signaled
-                              NULL); // anonymous
-    }
-
-    ~WEvent()
-    {
-        ::CloseHandle(event);
-    }
-
-    void    signal()
-    {
-        ::SetEvent(event);
-    }
-
-    void    wait(int timeout = 30000)
-    {
-        switch(::WaitForSingleObject(event, timeout))
-        {
-        case WAIT_TIMEOUT:
-            throw TimeoutException();
-            break;
-        //case WAIT_OBJECT_0:
-            //break;
-        }
-    }
-
-    void    reset()
-    {
-        ::ResetEvent(event);
-    }
-
-    HANDLE event;
-};
-#endif
-
 #ifdef _UNIX
-// ------------------------------------
 #include <errno.h>
 
 #ifdef __APPLE__
 #include <sched.h>
 #define _BIG_ENDIAN 1
 #endif
-
-// ------------------------------------
-class WEvent
-{
-public:
-
-    WEvent()
-    {
-    }
-
-    void    signal()
-    {
-    }
-
-    void    wait(int timeout = 30000)
-    {
-    }
-
-    void    reset()
-    {
-    }
-};
 #endif
 
-// ------------------------------------
-class WLock
-{
-private:
-    std::recursive_mutex m_mutex;
-
-public:
-    void    on()
-    {
-        m_mutex.lock();
-    }
-
-    void    off()
-    {
-        m_mutex.unlock();
-    }
-};
-
-// ------------------------------------
-typedef int (*THREAD_FUNC)(ThreadInfo *);
-#define THREAD_PROC int
-typedef std::thread THREAD_HANDLE;
-
-class ThreadInfo
-{
-public:
-    ThreadInfo()
-        : m_active(false)
-    {
-        func         = NULL;
-        data         = NULL;
-    }
-
-    void    shutdown();
-
-    bool active() { return m_active.load(); }
-
-    std::atomic<bool>   m_active;
-    THREAD_FUNC     func;
-    void            *data;
-
-    THREAD_HANDLE   handle;
-};
-
-// ------------------------------------
-class LogBuffer
-{
-public:
-    enum TYPE
-    {
-        T_NONE,
-        T_DEBUG,
-        T_ERROR,
-        T_NETWORK,
-        T_CHANNEL,
-    };
-
-    LogBuffer(int i, int l)
-    {
-        lineLen = l;
-        maxLines = i;
-        currLine = 0;
-        buf = new char[lineLen*maxLines];
-        times = new unsigned int [maxLines];
-        types = new TYPE [maxLines];
-    }
-
-    ~LogBuffer()
-    {
-        delete[] buf;
-        delete[] times;
-        delete[] types;
-    }
-
-    void    clear()
-    {
-        currLine = 0;
-    }
-
-    void                write(const char *, TYPE);
-    static const char   *getTypeStr(TYPE t) { return logTypes[t]; }
-    void                dumpHTML(class Stream &);
-
-    static void         escapeHTML(char* dest, char* src);
-
-    char            *buf;
-    unsigned int    *times;
-    unsigned int    currLine, maxLines, lineLen;
-    TYPE            *types;
-    WLock           lock;
-    static          const char *logTypes[];
-};
-
-// ------------------------------------
-extern Sys *sys;
+#ifdef WIN32
+typedef __int64 int64_t;
+#endif
 
 // ------------------------------------
